+++ lib/experiments-page.php
@@ -52,17 +52,6 @@ function gutenberg_initialize_experiments_settings() {
 			'id'    => 'gutenberg-list-v2',
 		)
 	);
-	add_settings_field(
-		'gutenberg-quote-v2',
-		__( 'Quote block v2', 'gutenberg' ),
-		'gutenberg_display_experiment_field',
-		'gutenberg-experiments',
-		'gutenberg_experiments_section',
-		array(
-			'label' => __( 'Test a new quote block that allows nested blocks (Warning: The new block is not ready. You may experience content loss, avoid using it on production sites)', 'gutenberg' ),
-			'id'    => 'gutenberg-quote-v2',
-		)
-	);
 	register_setting(
 		'gutenberg-experiments',
 		'gutenberg-experiments'
