+++ packages/block-library/src/quote/edit.js
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import classnames from 'classnames';
+import classNames from 'classnames';
 
 /**
  * WordPress dependencies
@@ -12,29 +12,87 @@ import {
 	BlockControls,
 	RichText,
 	useBlockProps,
+	useInnerBlocksProps,
+	store as blockEditorStore,
 } from '@wordpress/block-editor';
 import { BlockQuotation } from '@wordpress/components';
+import { useDispatch, useSelect, useRegistry } from '@wordpress/data';
 import { createBlock, getDefaultBlockName } from '@wordpress/blocks';
-import { Platform } from '@wordpress/element';
+import { Platform, useEffect } from '@wordpress/element';
+import deprecated from '@wordpress/deprecated';
+
+/**
+ * Internal dependencies
+ */
+import { migrateToQuoteV2 } from './deprecated';
 
 const isWebPlatform = Platform.OS === 'web';
+
+const TEMPLATE = [ [ 'core/paragraph', {} ] ];
+
+/**
+ * At the moment, deprecations don't handle create blocks from attributes
+ * (like when using CPT templates). For this reason, this hook is necessary
+ * to avoid breaking templates using the old quote block format.
+ *
+ * @param {Object} attributes Block attributes.
+ * @param {string} clientId   Block client ID.
+ */
+const useMigrateOnLoad = ( attributes, clientId ) => {
+	const registry = useRegistry();
+	const { updateBlockAttributes, replaceInnerBlocks } =
+		useDispatch( blockEditorStore );
+	useEffect( () => {
+		// As soon as the block is loaded, migrate it to the new version.
+
+		if ( ! attributes.value ) {
+			// No need to migrate if it doesn't have the value attribute.
+			return;
+		}
+
+		const [ newAttributes, newInnerBlocks ] =
+			migrateToQuoteV2( attributes );
+
+		deprecated( 'Value attribute on the quote block', {
+			since: '6.0',
+			version: '6.5',
+			alternative: 'inner blocks',
+		} );
+
+		registry.batch( () => {
+			updateBlockAttributes( clientId, newAttributes );
+			replaceInnerBlocks( clientId, newInnerBlocks );
+		} );
+	}, [ attributes.value ] );
+};
 
 export default function QuoteEdit( {
 	attributes,
 	setAttributes,
 	isSelected,
-	mergeBlocks,
-	onReplace,
-	className,
 	insertBlocksAfter,
+	clientId,
+	className,
 	style,
 } ) {
-	const { align, value, citation } = attributes;
+	const { align, citation } = attributes;
+
+	useMigrateOnLoad( attributes, clientId );
+
+	const isAncestorOfSelectedBlock = useSelect( ( select ) =>
+		select( blockEditorStore ).hasSelectedInnerBlock( clientId )
+	);
+	const hasSelection = isSelected || isAncestorOfSelectedBlock;
+
 	const blockProps = useBlockProps( {
-		className: classnames( className, {
+		className: classNames( className, {
 			[ `has-text-align-${ align }` ]: align,
 		} ),
-		style,
+		...( ! isWebPlatform && { style } ),
+	} );
+	const innerBlocksProps = useInnerBlocksProps( blockProps, {
+		template: TEMPLATE,
+		templateInsertUpdatesSelection: true,
 	} );
 
 	return (
@@ -47,66 +105,33 @@ export default function QuoteEdit( {
 					} }
 				/>
 			</BlockControls>
-			<BlockQuotation { ...blockProps }>
-				<RichText
-					identifier="value"
-					multiline
-					value={ value }
-					onChange={ ( nextValue ) =>
-						setAttributes( {
-							value: nextValue,
-						} )
-					}
-					onMerge={ mergeBlocks }
-					onRemove={ ( forward ) => {
-						const hasEmptyCitation =
-							! citation || citation.length === 0;
-						if ( ! forward && hasEmptyCitation ) {
-							onReplace( [] );
-						}
-					} }
-					aria-label={ __( 'Quote text' ) }
-					placeholder={
-						// translators: placeholder text used for the quote
-						__( 'Add quote' )
-					}
-					onReplace={ onReplace }
-					onSplit={ ( piece ) =>
-						createBlock( 'core/quote', {
-							...attributes,
-							value: piece,
-						} )
-					}
-					__unstableOnSplitMiddle={ () =>
-						createBlock( 'core/paragraph' )
-					}
-					textAlign={ align }
-					__unstableEmbedURLOnPaste
-				/>
-				{ ( ! RichText.isEmpty( citation ) || isSelected ) && (
+			<BlockQuotation { ...innerBlocksProps }>
+				{ innerBlocksProps.children }
+				{ ( ! RichText.isEmpty( citation ) || hasSelection ) && (
 					<RichText
 						identifier="citation"
 						tagName={ isWebPlatform ? 'cite' : undefined }
 						style={ { display: 'block' } }
 						value={ citation }
-						onChange={ ( nextCitation ) =>
+						onChange={ ( nextCitation ) => {
 							setAttributes( {
 								citation: nextCitation,
-							} )
-						}
+							} );
+						} }
 						__unstableMobileNoFocusOnMount
-						aria-label={ __( 'Quote citation text' ) }
+						aria-label={ __( 'Quote citation' ) }
 						placeholder={
-							// translators: placeholder text used for the citation
+							// translators: placeholder text used for the
+							// citation
 							__( 'Add citation' )
 						}
 						className="wp-block-quote__citation"
-						textAlign={ align }
 						__unstableOnSplitAtEnd={ () =>
 							insertBlocksAfter(
 								createBlock( getDefaultBlockName() )
 							)
 						}
+						{ ...( ! isWebPlatform ? { textAlign: align } : {} ) }
 					/>
 				) }
 			</BlockQuotation>
