+++ packages/e2e-tests/specs/editor/blocks/quote.test.js
@@ -77,7 +77,9 @@ describe( 'Quote', () => {
 			await page.keyboard.type( 'one' );
 			await page.keyboard.press( 'Enter' );
 			await page.keyboard.type( 'two' );
-			await transformBlockTo( 'Paragraph' );
+			// Navigate to the citation to select the block.
+			await page.keyboard.press( 'ArrowRight' );
+			await transformBlockTo( 'Unwrap' );
 
 			expect( await getEditedPostContent() ).toMatchSnapshot();
 		} );
@@ -89,7 +91,7 @@ describe( 'Quote', () => {
 			await page.keyboard.type( 'two' );
 			await page.keyboard.press( 'ArrowRight' );
 			await page.keyboard.type( 'cite' );
-			await transformBlockTo( 'Paragraph' );
+			await transformBlockTo( 'Unwrap' );
 
 			expect( await getEditedPostContent() ).toMatchSnapshot();
 		} );
@@ -98,14 +100,15 @@ describe( 'Quote', () => {
 			await insertBlock( 'Quote' );
 			await page.keyboard.press( 'ArrowRight' );
 			await page.keyboard.type( 'cite' );
-			await transformBlockTo( 'Paragraph' );
+			await transformBlockTo( 'Unwrap' );
 
 			expect( await getEditedPostContent() ).toMatchSnapshot();
 		} );
 
 		it( 'and renders a void paragraph if both the cite and quote are void', async () => {
 			await insertBlock( 'Quote' );
-			await transformBlockTo( 'Paragraph' );
+			await page.keyboard.press( 'ArrowRight' ); // Select the quote
+			await transformBlockTo( 'Unwrap' );
 
 			expect( await getEditedPostContent() ).toMatchSnapshot();
 		} );
@@ -116,56 +119,6 @@ describe( 'Quote', () => {
 		await page.keyboard.type( 'test' );
 		await transformBlockTo( 'Quote' );
 
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-	} );
-
-	it( 'is transformed to an empty heading if the quote is empty', async () => {
-		await insertBlock( 'Quote' );
-		await transformBlockTo( 'Heading' );
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-	} );
-
-	it( 'is transformed to a heading if the quote just contains one paragraph', async () => {
-		await insertBlock( 'Quote' );
-		await page.keyboard.type( 'one' );
-		await transformBlockTo( 'Heading' );
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-	} );
-
-	it( 'is transformed to a heading and a quote if the quote contains multiple paragraphs', async () => {
-		await insertBlock( 'Quote' );
-		await page.keyboard.type( 'one' );
-		await page.keyboard.press( 'Enter' );
-		await page.keyboard.type( 'two' );
-		await page.keyboard.press( 'Enter' );
-		await page.keyboard.type( 'three' );
-		await transformBlockTo( 'Heading' );
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-	} );
-
-	it( 'is transformed to a heading and a quote if the quote contains a citation', async () => {
-		await insertBlock( 'Quote' );
-		await page.keyboard.type( 'one' );
-		await page.keyboard.press( 'ArrowRight' );
-		await page.keyboard.type( 'cite' );
-		await transformBlockTo( 'Heading' );
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-	} );
-
-	it( 'the resuling quote after transforming to a heading can be transformed again', async () => {
-		await insertBlock( 'Quote' );
-		await page.keyboard.type( 'one' );
-		await page.keyboard.press( 'Enter' );
-		await page.keyboard.type( 'two' );
-		await page.keyboard.press( 'ArrowRight' );
-		await page.keyboard.type( 'cite' );
-		await transformBlockTo( 'Heading' );
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-		await page.click( '[data-type="core/quote"]' );
-		await transformBlockTo( 'Heading' );
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-		await page.click( '[data-type="core/quote"]' );
-		await transformBlockTo( 'Heading' );
 		expect( await getEditedPostContent() ).toMatchSnapshot();
 	} );
 
@@ -180,16 +133,7 @@ describe( 'Quote', () => {
 		expect( await getEditedPostContent() ).toMatchSnapshot();
 	} );
 
-	it( 'can be merged into from a paragraph', async () => {
-		await insertBlock( 'Quote' );
-		await insertBlock( 'Paragraph' );
-		await page.keyboard.type( 'test' );
-		await pressKeyTimes( 'ArrowLeft', 'test'.length );
-		await page.keyboard.press( 'Backspace' );
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-	} );
-
-	it( 'can be split at the end and merged back', async () => {
+	it( 'can be split at the end', async () => {
 		await insertBlock( 'Quote' );
 		await page.keyboard.type( '1' );
 		await page.keyboard.press( 'Enter' );
@@ -200,46 +144,7 @@ describe( 'Quote', () => {
 
 		await page.keyboard.press( 'Backspace' );
 
-		// Expect empty paragraph inside quote block.
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-
-		await page.keyboard.press( 'Backspace' );
-
-		// Expect quote without empty paragraphs.
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-	} );
-
-	it( 'can be split in the middle and merged back', async () => {
-		await insertBlock( 'Quote' );
-		await page.keyboard.type( '1' );
-		await page.keyboard.press( 'Enter' );
-		await page.keyboard.type( '2' );
-		await page.keyboard.press( 'ArrowRight' );
-		await page.keyboard.type( 'c' );
-		await page.keyboard.press( 'ArrowUp' );
-		await page.keyboard.press( 'ArrowUp' );
-		await page.keyboard.press( 'Enter' );
-		await page.keyboard.press( 'Enter' );
-
-		// Expect two quote blocks and empty paragraph in the middle.
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-
-		await page.keyboard.press( 'Backspace' );
-
-		// Expect two quote blocks and empty paragraph in the first quote.
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-
-		await page.keyboard.press( 'Backspace' );
-
-		// Expect two quote blocks.
-		expect( await getEditedPostContent() ).toMatchSnapshot();
-
-		await page.keyboard.press( 'ArrowLeft' );
-		await page.keyboard.press( 'ArrowDown' );
-		await page.keyboard.press( 'ArrowDown' );
-		await page.keyboard.press( 'ArrowDown' );
-		await page.keyboard.press( 'Backspace' );
-
+		// Expect the paragraph to be deleted.
 		expect( await getEditedPostContent() ).toMatchSnapshot();
 	} );
 } );
