+++ packages/edit-post/src/test/editor.native.js
@@ -35,9 +35,7 @@ afterAll( () => {
 } );
 
 describe( 'Editor', () => {
-	beforeAll( () => {
-		registerCoreBlocks();
-	} );
+	beforeAll( registerCoreBlocks );
 
 	it( 'detects unsupported block and sends hasUnsupportedBlocks true to native', () => {
 		RNReactNativeGutenbergBridge.editorDidMount = jest.fn();
