+++ test/e2e/specs/editor/blocks/quote.spec.js
@@ -23,7 +23,9 @@ test.describe( 'adding a quote', () => {
 		const content = await editor.getEditedPostContent();
 		expect( content ).toBe(
 			`<!-- wp:quote -->
-<blockquote class="wp-block-quote"><p>Quote content</p></blockquote>
+<blockquote class="wp-block-quote"><!-- wp:paragraph -->
+<p>Quote content</p>
+<!-- /wp:paragraph --></blockquote>
 <!-- /wp:quote -->`
 		);
 	} );
