+++ test/e2e/specs/editor/various/style-variation.spec.js
@@ -17,6 +17,9 @@ test.describe( 'adding blocks', () => {
 			attributes: { value: '<p>Quote content</p>' },
 		} );
 
+		// Select the quote block.
+		await page.keyboard.press( 'ArrowDown' );
+
 		await editor.clickBlockToolbarButton( 'Quote' );
 
 		await page.click( 'role=menuitem[name="Plain"i]' );
@@ -25,7 +28,9 @@ test.describe( 'adding blocks', () => {
 		const content = await editor.getEditedPostContent();
 		expect( content ).toBe(
 			`<!-- wp:quote {"className":"is-style-plain"} -->
-<blockquote class="wp-block-quote is-style-plain"><p>Quote content</p></blockquote>
+<blockquote class="wp-block-quote is-style-plain"><!-- wp:paragraph -->
+<p>Quote content</p>
+<!-- /wp:paragraph --></blockquote>
 <!-- /wp:quote -->`
 		);
 	} );
