+++ test/integration/fixtures/blocks/core__quote__deprecated-1.serialized.html
@@ -1,3 +1,5 @@
 <!-- wp:quote -->
-<blockquote class="wp-block-quote"><p>Testing deprecated quote block...</p><cite>...with a caption</cite></blockquote>
+<blockquote class="wp-block-quote"><!-- wp:paragraph -->
+<p>Testing deprecated quote block...</p>
+<!-- /wp:paragraph --><cite>...with a caption</cite></blockquote>
 <!-- /wp:quote -->
