+++ test/integration/fixtures/blocks/core__quote__deprecated-2.json
@@ -3,10 +3,19 @@
 		"name": "core/quote",
 		"isValid": true,
 		"attributes": {
-			"value": "<p>Testing deprecated quote block...</p>",
 			"citation": "...with a caption",
 			"align": "right"
 		},
-		"innerBlocks": []
+		"innerBlocks": [
+			{
+				"name": "core/paragraph",
+				"isValid": true,
+				"attributes": {
+					"content": "Testing deprecated quote block...",
+					"dropCap": false
+				},
+				"innerBlocks": []
+			}
+		]
 	}
 ]
