+++ test/integration/fixtures/blocks/core__quote__deprecated-2.serialized.html
@@ -1,3 +1,5 @@
 <!-- wp:quote {"align":"right"} -->
-<blockquote class="wp-block-quote has-text-align-right"><p>Testing deprecated quote block...</p><cite>...with a caption</cite></blockquote>
+<blockquote class="wp-block-quote has-text-align-right"><!-- wp:paragraph -->
+<p>Testing deprecated quote block...</p>
+<!-- /wp:paragraph --><cite>...with a caption</cite></blockquote>
 <!-- /wp:quote -->
