+++ test/integration/fixtures/blocks/core__quote__style-1.html
@@ -1,3 +1,5 @@
-<!-- wp:core/quote -->
-<blockquote class="wp-block-quote"><p>The editor will endeavour to create a new page and post building experience that makes writing rich posts effortless, and has “blocks” to make it easy what today might take shortcodes, custom HTML, or “mystery meat” embed discovery.</p><cite>Matt Mullenweg, 2017</cite></blockquote>
-<!-- /wp:core/quote -->
+<!-- wp:quote -->
+<blockquote class="wp-block-quote"><!-- wp:paragraph -->
+<p>The editor will endeavour to create a new page and post building experience that makes writing rich posts effortless, and has “blocks” to make it easy what today might take shortcodes, custom HTML, or “mystery meat” embed discovery.</p>
+<!-- /wp:paragraph --><cite>Matt Mullenweg, 2017</cite></blockquote>
+<!-- /wp:quote -->
