+++ test/integration/fixtures/blocks/core__quote__style-2.html
@@ -1,3 +1,5 @@
-<!-- wp:core/quote {"className":"is-style-large"} -->
-<blockquote class="wp-block-quote is-style-large"><p>There is no greater agony than bearing an untold story inside you.</p><cite>Maya Angelou</cite></blockquote>
-<!-- /wp:core/quote -->
+<!-- wp:quote {"className":"is-style-large"} -->
+<blockquote class="wp-block-quote is-style-large"><!-- wp:paragraph -->
+<p>There is no greater agony than bearing an untold story inside you.</p>
+<!-- /wp:paragraph --><cite>Maya Angelou</cite></blockquote>
+<!-- /wp:quote -->
