+++ test/integration/fixtures/blocks/core__quote__style-2.json
@@ -3,10 +3,20 @@
 		"name": "core/quote",
 		"isValid": true,
 		"attributes": {
-			"value": "<p>There is no greater agony than bearing an untold story inside you.</p>",
+			"value": "",
 			"citation": "Maya Angelou",
 			"className": "is-style-large"
 		},
-		"innerBlocks": []
+		"innerBlocks": [
+			{
+				"name": "core/paragraph",
+				"isValid": true,
+				"attributes": {
+					"content": "There is no greater agony than bearing an untold story inside you.",
+					"dropCap": false
+				},
+				"innerBlocks": []
+			}
+		]
 	}
 ]
