+++ test/integration/fixtures/blocks/core__quote__style-2.parsed.json
@@ -4,10 +4,22 @@
 		"attrs": {
 			"className": "is-style-large"
 		},
-		"innerBlocks": [],
-		"innerHTML": "\n<blockquote class=\"wp-block-quote is-style-large\"><p>There is no greater agony than bearing an untold story inside you.</p><cite>Maya Angelou</cite></blockquote>\n",
+		"innerBlocks": [
+			{
+				"blockName": "core/paragraph",
+				"attrs": {},
+				"innerBlocks": [],
+				"innerHTML": "\n<p>There is no greater agony than bearing an untold story inside you.</p>\n",
+				"innerContent": [
+					"\n<p>There is no greater agony than bearing an untold story inside you.</p>\n"
+				]
+			}
+		],
+		"innerHTML": "\n<blockquote class=\"wp-block-quote is-style-large\"><cite>Maya Angelou</cite></blockquote>\n",
 		"innerContent": [
-			"\n<blockquote class=\"wp-block-quote is-style-large\"><p>There is no greater agony than bearing an untold story inside you.</p><cite>Maya Angelou</cite></blockquote>\n"
+			"\n<blockquote class=\"wp-block-quote is-style-large\">",
+			null,
+			"<cite>Maya Angelou</cite></blockquote>\n"
 		]
 	}
 ]
