+++ test/integration/fixtures/documents/markdown-out.html
@@ -39,7 +39,13 @@
 <!-- /wp:heading -->
 
 <!-- wp:quote -->
-<blockquote class="wp-block-quote"><p>First</p><p>Second</p></blockquote>
+<blockquote class="wp-block-quote"><!-- wp:paragraph -->
+<p>First</p>
+<!-- /wp:paragraph -->
+
+<!-- wp:paragraph -->
+<p>Second</p>
+<!-- /wp:paragraph --></blockquote>
 <!-- /wp:quote -->
 
 <!-- wp:heading -->
